Hides tasks from the front page that are in finished projects.
This was done by adding the condition this.project.state == 'running' to the
view, where it checks if there are any items to do.
Note: There is a scope defined in the 'project.rb' model that makes a similar
comparison, but I could not apply the scope to a single object as it responded
with "Error: undefined method `active' for "Test":Project".
If changes are made to the states project has, then this may break the
front page task view.